% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Panel_report_decision.R
\name{check_panel_dataframe}
\alias{check_panel_dataframe}
\title{Check number of entries in panel dataframe, concordance in panel and discordance in indeces}
\usage{
check_panel_dataframe(panel_df)
}
\arguments{
\item{panel_df}{}
}
\description{
Check number of entries in panel dataframe, concordance in panel and discordance in indeces
}
