% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seq_panel.R
\name{check_seq_panel}
\alias{check_seq_panel}
\title{Check amplicon panel used in}
\usage{
check_seq_panel(file = clonesummary)
}
\arguments{
\item{clonesummary}{}
}
\description{
Check amplicon panel used in
}
