% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clone_frequencies_clonsum}
\alias{clone_frequencies_clonsum}
\title{Plot clone frequencies vs productive/unproductive category}
\usage{
clone_frequencies_clonsum(plot_prepped)
}
\arguments{
\item{plot_prepped}{}
}
\description{
Plot clone frequencies vs productive/unproductive category
}
