% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clone_frequencies_lineage}
\alias{clone_frequencies_lineage}
\title{Plot lineage frequencies vs productive/unproductive category}
\usage{
clone_frequencies_lineage(plot_prepped)
}
\arguments{
\item{plot_prepped}{}
}
\description{
Plot lineage frequencies vs productive/unproductive category
}
