% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clone_frequencies_vs_rank_clonesum}
\alias{clone_frequencies_vs_rank_clonesum}
\title{Plot clone frequencies vs rank -- clone summary file}
\usage{
clone_frequencies_vs_rank_clonesum(plot_prepped)
}
\arguments{
\item{plot_prepped}{}
}
\description{
Plot clone frequencies vs rank -- clone summary file
}
