% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clone_frequencies_vs_rank_lineage}
\alias{clone_frequencies_vs_rank_lineage}
\title{Plot clone frequencies vs rank -- lineage summary file}
\usage{
clone_frequencies_vs_rank_lineage(plot_prepped)
}
\arguments{
\item{plot_prepped}{}
}
\description{
Plot clone frequencies vs rank -- lineage summary file
}
