% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonesummaries_read.R
\name{clonesum_read}
\alias{clonesum_read}
\title{Read  clone summary file}
\usage{
clonesum_read(inputpath)
}
\arguments{
\item{filepath}{}
}
\value{
Read in clone summary with filepath as additional column
}
\description{
Read  clone summary file
}
