% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seq_panel.R
\name{clonesummary_indexing}
\alias{clonesummary_indexing}
\title{Indexing function to assign order for input files}
\usage{
clonesummary_indexing(charstring = check_seq_panel_output)
}
\arguments{
\item{check_seq_panel_output}{}
}
\description{
Indexing function to assign order for input files
}
