% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seq_panel.R
\name{clonesummary_panel}
\alias{clonesummary_panel}
\title{Indexing function to assign order for input files}
\usage{
clonesummary_panel(charstring = check_seq_panel_output)
}
\arguments{
\item{check_seq_panel_output}{}
}
\description{
Indexing function to assign order for input files
}
