% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_extract.R
\name{filename_extract}
\alias{filename_extract}
\title{Extract common string from filenames}
\usage{
filename_extract(filestr1, filestr2)
}
\arguments{
\item{filevector}{}
}
\value{
substring present in both filestrings
}
\description{
Extract common string from filenames
}
\examples{
banana AND ananas -> anana
will fail with more than two strings
}
