% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_clonesummaries.R
\name{list_clonesummaries}
\alias{list_clonesummaries}
\title{List only clonesummaries in directory}
\usage{
list_clonesummaries(filedir)
}
\arguments{
\item{filedir}{}
}
\value{
Return filevector of clone summaries actively excluding 'downsample' clonesummaries
}
\description{
List only clonesummaries in directory
}
