% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_lineage_summaries.R
\name{list_lineage_summaries}
\alias{list_lineage_summaries}
\title{List only lineage summaries in directory}
\usage{
list_lineage_summaries(filedir)
}
\arguments{
\item{filedir}{}
}
\value{
Return filevector of lineage summaries actively excluding 'downsample' files
}
\description{
List only lineage summaries in directory
}
