% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_clonesummaries.R
\name{only_clonesummaries}
\alias{only_clonesummaries}
\title{List only clonesummaries in directory}
\usage{
only_clonesummaries(filedir)
}
\arguments{
\item{filedir}{}
}
\description{
List only clonesummaries in directory
}
