% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_background.R
\name{overbackground_bcr}
\alias{overbackground_bcr}
\title{Determine if largest clone frequency is above background}
\usage{
overbackground_bcr(clonesum)
}
\arguments{
\item{clonesum}{}
}
\description{
Determine if largest clone frequency is above background
}
