% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_background.R
\name{overbackground_tcr}
\alias{overbackground_tcr}
\title{Determine if largest clone frequency is above background}
\usage{
overbackground_tcr(clonesum)
}
\arguments{
\item{clonsum}{}
}
\description{
Determine if largest clone frequency is above background
}
