% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Panel_report_decision.R
\name{panel_dataframe}
\alias{panel_dataframe}
\title{Extract info from clone summary and convert to dataframe}
\usage{
panel_dataframe(panel_df)
}
\arguments{
\item{filevector}{}
}
\description{
Extract info from clone summary and convert to dataframe
}
