% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Panel_report_decision.R
\name{panel_decision}
\alias{panel_decision}
\title{Returns string for panel based on panel dataframe.
Will be used to trigger correct reporting in LymphoyteSeq_report render script.}
\usage{
panel_decision(panel_df)
}
\arguments{
\item{panel_df}{}
}
\value{
string
}
\description{
Returns string for panel based on panel dataframe.
Will be used to trigger correct reporting in LymphoyteSeq_report render script.
}
