% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Panel_report_decision.R
\name{path_to_info}
\alias{path_to_info}
\title{TCR/BCR report decision
Aggregate file info
From filepaths to dataframe}
\usage{
path_to_info(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
TCR/BCR report decision
Aggregate file info
From filepaths to dataframe
}
