% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topClone.R
\name{top25_clonesum_selection_bcr}
\alias{top25_clonesum_selection_bcr}
\title{Return top25 clones of clone summaries formatted for presentation in Data Table}
\usage{
top25_clonesum_selection_bcr(clonesum)
}
\arguments{
\item{clonesum}{}
}
\description{
Return top25 clones of clone summaries formatted for presentation in Data Table
}
