% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topClone.R
\name{top25_clonesum_selection_tcr}
\alias{top25_clonesum_selection_tcr}
\title{Return top25 clones of TCR cloneummaries formatted for presentation in Data Table}
\usage{
top25_clonesum_selection_tcr(clonesum)
}
\arguments{
\item{lineage_file}{}
}
\description{
Return top25 clones of TCR cloneummaries formatted for presentation in Data Table
}
