% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topClone.R
\name{top25_lineage_selection}
\alias{top25_lineage_selection}
\title{Return top25 clones of lineage summaries formatted for presentation in Data Table}
\usage{
top25_lineage_selection(lineage_file)
}
\arguments{
\item{lineage_file}{}
}
\description{
Return top25 clones of lineage summaries formatted for presentation in Data Table
}
