% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{top_ratio_to_x}
\alias{top_ratio_to_x}
\title{Calculate Top clone enrichment of jth clone}
\usage{
top_ratio_to_x(clonesum, j)
}
\arguments{
\item{j}{}
}
\description{
Calculate Top clone enrichment of jth clone
}
