% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topClone.R
\name{topClone_analysis_clonesum}
\alias{topClone_analysis_clonesum}
\title{Retrieve the division result of i and jth clone frequencies}
\usage{
topClone_analysis_clonesum(clonesum, i, j)
}
\arguments{
\item{clonesum}{}
}
\value{
Return Frequency and sequence of top clone
}
\description{
Retrieve the division result of i and jth clone frequencies
}
