% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topClone.R
\name{topClone_analysis_lineage}
\alias{topClone_analysis_lineage}
\title{Retrieve the division result of i and jth clone}
\usage{
topClone_analysis_lineage(lineage, i, j)
}
\arguments{
\item{clonesum}{}
}
\value{
Return Frequency and sequence of top clone
}
\description{
Retrieve the division result of i and jth clone
}
