% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_counts.R
\name{total_counts}
\alias{total_counts}
\title{Bring rank counts back}
\usage{
total_counts(clone_summary, rank)
}
\arguments{
\item{clone_summary}{}

\item{rank}{}
}
\description{
Bring rank counts back
}
