% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{total_read_counts}
\alias{total_read_counts}
\title{Plot total read counts with productive/unproductive separation}
\usage{
total_read_counts(plot_prepped)
}
\arguments{
\item{plot_prepped}{}
}
\description{
Plot total read counts with productive/unproductive separation
}
